legal-service/requirements.txt
# LangChain & LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0

# Weaviate
weaviate-client>=4.0.0

# OpenAI 
openai>=1.12.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.5.2

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv>=1.0.0
redis>=5.0.0

# Logging
structlog>=24.1.0

